some refactoring, renaming to GraphEditor
