+++ IncrGraph/backend/.gitignore
@@ -4,3 +4,4 @@ logs
 tsconfig.tsbuildinfo
 node_modules
 dist
+cache.json
