+++ IncrGraph/frontend/package.json
@@ -22,7 +22,6 @@
 		"@xterm/addon-fit": "^0.10.0",
 		"@xterm/xterm": "^5.5.0",
 		"axios": "^1.7.7",
-		"dom-to-svg": "^0.12.2",
 		"html-to-image": "^1.11.13",
 		"js-cookie": "^3.0.5",
 		"lodash": "^4.17.21",
