+++ IncrGraph/frontend/src/main.tsx
@@ -1,4 +1,3 @@
-import "../wdyr.js";
 import { createRoot } from "react-dom/client";
 import { RouterProvider } from "react-router-dom";
